+++ src/app/service/new-user/existing-email/existing-email.service.ts
@@ -1,7 +1,6 @@
 import { Injectable } from '@angular/core';
 import { AbstractControl } from '@angular/forms';
 import { first, map, switchMap } from 'rxjs/operators';
-
 import { NewUserService } from '../new-user.service';
 
 @Injectable({
