+++ src/app/service/user/user.d.ts
@@ -1,4 +1,4 @@
-export interface Auth {
+export interface User {
   id?: number;
   name?: string;
   email?: string;
